split into routes & pages
