+++ src/pages/StartMenu.tsx
@@ -1,0 +1,3 @@
+export const StartMenu = () => {
+  return <div>TODO: START MENU</div>
+}
